chore(orchestrate): refactor files
