+++ orchestrate/tests/crypto_verify.rs
@@ -1,6 +1,6 @@
 use cosmwasm_orchestrate::{
-    vm::{Account, State, StateBuilder},
-    Api, Unit,
+    vm::{Account, State},
+    Api, StateBuilder, Unit,
 };
 use cosmwasm_std::{from_binary, Binary, BlockInfo, ContractInfo, Env, MessageInfo, Timestamp};
 use serde::{Deserialize, Serialize};
